* DIPK fix for batch size 1 * MOLIR/SuperFELTR fix for if self.model is None * MOLIR/SuperFELTR: moved duplicated code to utils * MultiOmicsNN: removed unused self.methylation_features * test restructuring: now baseline tests, single drug model tests, neural network tests for non-single drug models * Also adapted tests for the bugs
